uievents/order-of-events/mouse-events/wheel-basic-manual.html

<!doctype html>
<head>
  <meta charset=utf-8>
  <title>WheelEvent - basic wheel event</title>
  <style>
    .testarea{ margin: auto; width: 80%; height: 250px; border: 1px solid grey; position: relative; }

    #wheelbox { background-color: red; border: 1px solid black; margin: 0; padding: 0; }
    #wheelbox.green { background-color: green; }
    #wheelbox { position: absolute; left: 15%; top: 15%; width: 50%; height: 50% }
  </style>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    setup({explicit_timeout: true});
  </script>
  <script src="/uievents/resources/eventrecorder.js"></script>
</head>

<body>
  <p><strong>Description</strong>: Verifies that wheel events are captured and sent</p>
  <p><strong>Instructions</strong>: </p>
  <ol>
    <li>Place your mouse over the box</li>
    <li>Scroll the mouse wheel to change the box color</li>
  </ol>
  <p><strong>Test Passes</strong> if the box turns green and the word 'PASS' appears below</p>

  <section class="testarea">
    <div id="wheelbox"></div>
  </section>

  <script>
    var wheelbox = document.getElementById("wheelbox");

    EventRecorder.configure({
      mergeEventTypes: ['wheel'],
      objectMap: {
        "div#wheelbox": wheelbox
      }
    });

    wheelbox.addRecordedEventListener('wheel', function (e) {
      e.stopPropagation();
      this.className = "green";
      endTest();
      done();
    });

    function endTest() {
      EventRecorder.stop();
      var results = EventRecorder.getRecords();
      // Check results:
      assert_equals(results.length, 1, "Two mousemove events");
      assert_equals(results[0].event.type, 'wheel', "First event is a wheel event");
      assert_equals(results[0].event.target, 'div#wheelbox', "First event targetted wheelbox");
    }

    EventRecorder.start();
  </script>
</body>
</html>
